# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nimiq Mini Apps API",
    "version": "1.0.0",
    "description": "REST API for the [Nimiq Mini Apps](https://nimiqminiapps.com) community directory.\n\n**Submitting an app** — `POST /api/apps/submit` (wallet login required). Rate limit: **5 requests per hour per IP**.\nRequired field `submitter_contact` (Telegram, email, etc.) is stored privately and omitted from public responses.\n\n**Open in Nimiq Pay** — listed apps expose `open_url` as `https://nimpay.app/miniapps/open/{domain}` (domain without `https://`).\n\nMachine-readable spec: `GET /openapi.json` or `GET /openapi.yaml`.\n",
    "contact": {
      "name": "Nimiq Mini Apps",
      "url": "https://nimiqminiapps.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/NimMiniApps/NimiqMiniApps/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "https://api.nimiqminiapps.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/apps/submit": {
      "post": {
        "tags": [
          "Submit"
        ],
        "summary": "Submit a new app for review",
        "description": "Creates an app with `status=submitted` and `featured=false`. Requires a wallet\nsession (see `POST /api/auth/verify`); the caller must have a `display_name` set\non their profile first. `developer_slug`/`developer_name` are derived from the\ncaller's wallet identity, not taken from the request body.\nRate limited to 5 submissions per hour per client IP (HTTP 429).\n",
        "operationId": "submitApp",
        "security": [
          {
            "walletCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSubmitRequest"
              },
              "example": {
                "slug": "my-mini-app",
                "name": "My Mini App",
                "domain": "myapp.example.com",
                "category": "Games",
                "tagline": "One sentence pitch",
                "submitter_contact": "@telegram or you@example.com",
                "description": "Short plain-text summary for listings",
                "long_description": "## Features\n\n- **Bold** and lists work here",
                "tags": [
                  "games"
                ],
                "assets": [
                  "NIM"
                ],
                "reward_assets": [
                  "NIM"
                ],
                "release_stage": "beta"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "App submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppPublic"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "description": "Slug already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
    "/api/apps/{slug}/request-update": {
      "post": {
        "tags": [
          "Submit"
        ],
        "summary": "Request a listing update (reviewed by admins)",
        "description": "Proposes changes to an existing public listing. Requires a wallet session for one\nof the app's current owners (see `owner_wallet_addresses`). `developer_name`/`developer_slug` in the\nrequest body are ignored — identity changes are admin-only. Rate limited (same\n5/hour per IP as submit). Only one pending revision per app at a time.\n",
        "operationId": "requestAppUpdate",
        "security": [
          {
            "walletCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Update request queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateRequestCreated"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "The wallet session does not own this app",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "A pending update already exists for this app",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
    "/api/admin/revisions": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List pending update requests",
        "operationId": "adminListRevisions",
        "security": [
          {
            "adminBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "pending"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revisions with current app snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RevisionReviewItem"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Set ADMIN_TOKEN on the server; send as Authorization Bearer token"
      },
      "walletCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "wallet_session",
        "description": "Signed wallet-session cookie set by POST /api/auth/verify"
      }
    },
    "parameters": {
      "slug": {
        "name": "slug",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Invalid or missing admin token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded (5 requests per hour per IP for submit and request-update)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "CategoryName": {
        "type": "string",
        "enum": [
          "Games",
          "Utilities",
          "Finance",
          "Maps",
          "Social",
          "Experiments"
        ]
      },
      "AppStatus": {
        "type": "string",
        "enum": [
          "submitted",
          "approved",
          "verified",
          "experimental",
          "rejected"
        ]
      },
      "ReleaseStage": {
        "type": "string",
        "enum": [
          "concept",
          "alpha",
          "beta",
          "released"
        ],
        "default": "released"
      },
      "Asset": {
        "type": "string",
        "enum": [
          "NIM",
          "USDT",
          "USDC",
          "BTC",
          "ETH"
        ]
      },
      "MediaItem": {
        "type": "object",
        "required": [
          "type",
          "url"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "image",
              "youtube"
            ]
          },
          "url": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "SocialLink": {
        "type": "object",
        "required": [
          "platform",
          "url"
        ],
        "properties": {
          "platform": {
            "type": "string",
            "enum": [
              "twitter",
              "discord",
              "telegram",
              "bluesky",
              "instagram",
              "youtube",
              "linkedin",
              "mastodon",
              "reddit",
              "tiktok"
            ]
          },
          "url": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "AppInput": {
        "type": "object",
        "required": [
          "slug",
          "name",
          "domain",
          "category",
          "developer_slug",
          "developer_name",
          "tagline"
        ],
        "properties": {
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
          },
          "name": {
            "type": "string"
          },
          "domain": {
            "type": "string",
            "description": "Hostname only — must not include `https://`"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryName"
          },
          "developer_slug": {
            "type": "string"
          },
          "developer_name": {
            "type": "string"
          },
          "tagline": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "description": "Plain-text short summary; defaults to tagline if omitted"
          },
          "long_description": {
            "type": "string",
            "description": "Markdown for the app detail page"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          },
          "reward_assets": {
            "type": "array",
            "description": "Assets users can actually receive from this app, such as daily rewards, leaderboard prizes, payouts, tips, faucets, or similar receive-side flows. Leave empty when the app merely accepts, displays, swaps, or supports a token.",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          },
          "status": {
            "$ref": "#/components/schemas/AppStatus"
          },
          "release_stage": {
            "$ref": "#/components/schemas/ReleaseStage"
          },
          "featured": {
            "type": "boolean"
          },
          "featured_order": {
            "type": "integer",
            "minimum": 0
          },
          "website_url": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "github_url": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "icon_url": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "banner_url": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaItem"
            }
          },
          "socials": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialLink"
            }
          },
          "submitter_contact": {
            "type": "string",
            "maxLength": 200,
            "description": "Private contact; required for public submit"
          }
        }
      },
      "AppSubmitRequest": {
        "type": "object",
        "required": [
          "slug",
          "name",
          "domain",
          "category",
          "tagline",
          "submitter_contact"
        ],
        "properties": {
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
          },
          "name": {
            "type": "string"
          },
          "domain": {
            "type": "string",
            "description": "Hostname only — must not include `https://`"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryName"
          },
          "tagline": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "description": "Plain-text short summary; defaults to tagline if omitted"
          },
          "long_description": {
            "type": "string",
            "description": "Markdown for the app detail page"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          },
          "reward_assets": {
            "type": "array",
            "description": "Assets users can actually receive from this app, such as daily rewards, leaderboard prizes, payouts, tips, faucets, or similar receive-side flows. Leave empty when the app merely accepts, displays, swaps, or supports a token.",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          },
          "release_stage": {
            "$ref": "#/components/schemas/ReleaseStage"
          },
          "website_url": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "github_url": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "icon_url": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "banner_url": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaItem"
            }
          },
          "socials": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialLink"
            }
          },
          "submitter_contact": {
            "type": "string",
            "maxLength": 200,
            "description": "Private contact; required for public submit"
          }
        }
      },
      "AppUpdateRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AppInput"
          }
        ]
      },
      "AppPublic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryName"
          },
          "developer_slug": {
            "type": "string"
          },
          "developer_name": {
            "type": "string"
          },
          "owner_wallet_addresses": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Wallets that can manage this listing (My apps, request-update). Empty when unclaimed."
          },
          "tagline": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "long_description": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          },
          "reward_assets": {
            "type": "array",
            "description": "Assets users can actually receive from this app, such as daily rewards, leaderboard prizes, payouts, tips, faucets, or similar receive-side flows. Leave empty when the app merely accepts, displays, swaps, or supports a token.",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          },
          "status": {
            "$ref": "#/components/schemas/AppStatus"
          },
          "release_stage": {
            "$ref": "#/components/schemas/ReleaseStage"
          },
          "featured": {
            "type": "boolean"
          },
          "featured_order": {
            "type": "integer"
          },
          "website_url": {
            "type": "string",
            "nullable": true
          },
          "github_url": {
            "type": "string",
            "nullable": true
          },
          "icon_url": {
            "type": "string",
            "nullable": true
          },
          "discovered_icon_url": {
            "type": "string",
            "nullable": true
          },
          "banner_url": {
            "type": "string",
            "nullable": true
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaItem"
            }
          },
          "socials": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialLink"
            }
          },
          "domain_reachable": {
            "type": "boolean",
            "nullable": true
          },
          "domain_checked_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "open_url": {
            "type": "string",
            "format": "uri",
            "example": "https://nimpay.app/miniapps/open/myapp.example.com"
          }
        }
      },
      "AppAdmin": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AppPublic"
          },
          {
            "type": "object",
            "properties": {
              "submitter_contact": {
                "type": "string"
              },
              "total_opens": {
                "type": "integer",
                "description": "Lifetime opens (admin listing only)"
              },
              "total_views": {
                "type": "integer",
                "description": "Lifetime views (admin listing only)"
              }
            }
          }
        ]
      },
...
      "UpdateRequestCreated": {
        "type": "object",
        "properties": {
          "revision_id": {
            "type": "string",
            "format": "uuid"
          },
          "app_slug": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "example": "pending"
          }
        }
      },
...
      "RevisionReviewItem": {
        "type": "object",
        "properties": {
          "revision": {
            "$ref": "#/components/schemas/AppRevision"
          },
          "current": {
            "$ref": "#/components/schemas/AppAdmin"
          }
        }
      }
    }
  }
}